feat(teslaTower, Zombies): add new spries
