+++ packages/client/src/game/classes/gameEntities/Enemies/index.ts
@@ -1,7 +1,11 @@
 import ShieldZombie from "@/game/classes/gameEntities/Enemies/ShieldZombie";
 import ConeZombie from "@/game/classes/gameEntities/Enemies/coneZombie";
+import SaucepanZombie from "@/game/classes/gameEntities/Enemies/SaucepanZombie";
+import HelmetZombie from "@/game/classes/gameEntities/Enemies/HelmetZombie";
 
 export default {
   ShieldZombie,
   ConeZombie,
+  SaucepanZombie,
+  HelmetZombie,
 };
