+++ packages/client/src/game/mocks/level/index.ts
@@ -5,18 +5,20 @@ const dummyLevel = {
     // Wave 1
     {
       ShieldZombie: 5,
-      ConeZombie: 3,
+      ConeZombie: 4,
+      // HelmetZombie: 10,
     },
 
     // Wave 2
     {
-      ShieldZombie: 7,
+      ShieldZombie: 8,
       ConeZombie: 5,
+      // SaucepanZombie: 8,
     },
 
     // Wave 3ˇ
     {
-      ShieldZombie: 10,
+      ShieldZombie: 12,
       ConeZombie: 8,
     },
   ],
